+++ src/Common/Builder.fs
@@ -293,6 +293,7 @@ module ApplicationBuilder =
                 //
                 let preparedRuntimeParts: PreparedConsumeRuntimeParts<'OutputEvent> = {
                     Logger = logger
+                    IncrementMetric = ApplicationMetrics.incrementCustomMetricCount instance
                     Box = box
                     Environment = environment
                     Connections = connections