+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `IncrementMetric` to Consume Runtime Parts to allow custom metrics
 
 ## 1.0.0 - 2019-05-29
 - Initial implementation